# Codigo/backend/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "dtiflix-backend",
    "description": "Full stack application built with Next.js + TypeScript and Sequelize, integrating the OMDb API to search movies, manage favorites, write reviews, and display personalized statistics in a dashboard. Includes search history, average ratings, favorite genres/decades, and a responsive UI.",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    },
    "termsOfService": "http://localhost:5000/api/",
    "contact": {
      "name": "Davi Cândido",
      "email": "davicandidopucminas@gmail.com",
      "phone": "+55 (31) 97306-7259"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/",
      "description": "Ambiente de desenvolvimento"
    },
    {
      "url": "http://localhost:5000/api/",
      "description": "Ambiente de produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Favorite": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID do favorito"
          },
          "movie_id": {
            "type": "integer",
            "description": "ID do filme favoritado"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/favorites/search/{imdbID}": {
      "get": {
        "summary": "Busca um favorito pelo IMDb ID do filme",
        "tags": ["Favorites"],
        "parameters": [
          {
            "in": "path",
            "name": "imdbID",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "IMDb ID do filme"
          }
        ],
        "responses": {
          "200": {
            "description": "Favorito encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Favorite"
                }
              }
            }
          },
          "404": {
            "description": "Favorito não encontrado"
          }
        }
      }
    },
...
    "/favorites": {
      "get": {
        "summary": "Lista todos os favoritos",
        "tags": ["Favorites"],
        "responses": {
          "200": {
            "description": "Lista de favoritos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Favorite"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria um favorito",
        "tags": ["Favorites"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "description": "ID do filme"
                  }
                },
                "required": ["id"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Favorito criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Favorite"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deleta todos os favoritos",
        "tags": ["Favorites"],
        "responses": {
          "200": {
            "description": "Todos os favoritos deletados"
          }
        }
      }
    },
    "/favorites/{id}": {
      "get": {
        "summary": "Busca um favorito pelo ID",
        "tags": ["Favorites"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID do favorito"
          }
        ],
        "responses": {
          "200": {
            "description": "Favorito encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Favorite"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualiza um favorito",
        "tags": ["Favorites"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID do favorito"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Favorito atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Favorite"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deleta um favorito pelo ID",
        "tags": ["Favorites"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID do favorito"
          }
        ],
        "responses": {
          "200": {
            "description": "Favorito deletado"
          }
        }
      }
    },
...
  },
...
}